+++ docs/changelog.md
@@ -1,3 +1,3 @@
 ## v0.1.0
 
-First release of `box`.
+First release of `boxer`.
