+++ docs/cli.md
@@ -1,13 +1,13 @@
 # CLI documentation
 
-Below is the automatically generated docummentation for `box`.
+Below is the automatically generated docummentation for `boxer`.
 You can also find this by typing `box --help` in your terminal.
 
 
 
 ::: mkdocs-click
-    :module: box.cli
+    :module: boxer.cli
     :command: cli
-    :prog_name: box
+    :prog_name: boxer
     :depth: 1
     :style: table
