+++ tests/cli/test_boxer.py
@@ -4,7 +4,7 @@ import importlib.metadata
 
 from click.testing import CliRunner
 
-from box.cli import cli
+from boxer.cli import cli
 
 
 def test_version():
@@ -13,4 +13,4 @@ def test_version():
         result = runner.invoke(cli, ["--version"])
         assert result.exit_code == 0
         assert result.output.startswith("cli, version ")
-        assert result.output.rstrip().endswith(importlib.metadata.version("box"))
+        assert result.output.rstrip().endswith(importlib.metadata.version("boxer"))
