+++ tests/conftest.py
@@ -6,7 +6,7 @@ import subprocess
 
 import pytest
 
-from box.config import pyproject_writer
+from boxer.config import pyproject_writer
 
 
 @pytest.fixture
@@ -44,11 +44,11 @@ def tmp_path_chdir(tmp_path):
 
 @pytest.fixture
 def rye_project(tmp_path):
-    """Create a valid project for a minimum `rye` configured project with box."""
+    """Create a valid project for a minimum `rye` configured project with boxer."""
     # dump environment variables with subprocess
     subprocess.run("rye init .", cwd=tmp_path, shell=True, stdout=subprocess.DEVNULL)
 
-    # add builder for box to pyproject.toml
+    # add builder for boxer to pyproject.toml
     current_dir = Path().absolute()
     os.chdir(tmp_path)
     pyproject_writer("builder", "rye")
@@ -69,7 +69,7 @@ def rye_project_no_box(tmp_path):
     # dump environment variables with subprocess
     subprocess.run("rye init .", cwd=tmp_path, shell=True, stdout=subprocess.DEVNULL)
 
-    # add builder for box to pyproject.toml
+    # add builder for boxer to pyproject.toml
     current_dir = Path().absolute()
     os.chdir(tmp_path)
